Move readProgress into non-cmd package
