+++ cmd/analyze.go
@@ -3,9 +3,9 @@ package cmd
 import (
 	"log"
 	. "rare/cmd/helpers" //lint:ignore ST1001 Legacy
-	"rare/cmd/readProgress"
 	"rare/pkg/aggregation"
 	"rare/pkg/color"
+	"rare/pkg/extractor/readState"
 	"rare/pkg/humanize"
 	"rare/pkg/multiterm"
 	"strconv"
@@ -68,7 +68,7 @@ func analyzeFunction(c *cli.Context) error {
 	RunAggregationLoop(ext, aggr, func() {
 		line := writeAggrOutput(writer, aggr, extra, quantiles)
 		writer.WriteForLine(line+1, FWriteExtractorSummary(ext, aggr.ParseErrors()))
-		writer.WriteForLine(line+2, readProgress.GetReadFileString())
+		writer.WriteForLine(line+2, readState.GetReadFileString())
 	})
 
 	writer.Close()
