+++ cmd/histo.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"os"
 	. "rare/cmd/helpers" //lint:ignore ST1001 Legacy
-	"rare/cmd/readProgress"
 	"rare/pkg/aggregation"
 	"rare/pkg/color"
+	"rare/pkg/extractor/readState"
 	"rare/pkg/multiterm"
 	"rare/pkg/multiterm/termrenderers"
 
@@ -57,7 +57,7 @@ func histoFunction(c *cli.Context) error {
 	RunAggregationLoop(ext, counter, func() {
 		writeHistoOutput(writer, counter, topItems, reverseSort, sortByKey, atLeast)
 		writer.InnerWriter().WriteForLine(topItems, progressString())
-		writer.InnerWriter().WriteForLine(topItems+1, readProgress.GetReadFileString())
+		writer.InnerWriter().WriteForLine(topItems+1, readState.GetReadFileString())
 	})
 
 	writer.InnerWriter().Close()
