+++ cmd/tabulate.go
@@ -3,9 +3,9 @@ package cmd
 import (
 	"fmt"
 	. "rare/cmd/helpers" //lint:ignore ST1001 Legacy
-	"rare/cmd/readProgress"
 	"rare/pkg/aggregation"
 	"rare/pkg/color"
+	"rare/pkg/extractor/readState"
 	"rare/pkg/humanize"
 	"rare/pkg/multiterm"
 	"rare/pkg/multiterm/termrenderers"
@@ -56,7 +56,7 @@ func tabulateFunction(c *cli.Context) error {
 		}
 		writer.InnerWriter().WriteForLine(line, FWriteExtractorSummary(ext, counter.ParseErrors(),
 			fmt.Sprintf("(R: %v; C: %v)", color.Wrapi(color.Yellow, counter.RowCount()), color.Wrapi(color.BrightBlue, counter.ColumnCount()))))
-		writer.InnerWriter().WriteForLine(line+1, readProgress.GetReadFileString())
+		writer.InnerWriter().WriteForLine(line+1, readState.GetReadFileString())
 	})
 
 	writer.InnerWriter().Close()
