+++ pkg/extractor/readState/readProgress.go
@@ -1,4 +1,4 @@
-package readProgress
+package readState
 
 import (
 	"fmt"
