+++ pkg/extractor/readState/readProgress_test.go
@@ -1,4 +1,4 @@
-package readProgress
+package readState
 
 import (
 	"testing"
