+++ pkg/extractor/readState/util.go
@@ -1,4 +1,4 @@
-package readProgress
+package readState
 
 func min(a, b int) int {
 	if a < b {
